requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
PyJWT==2.10.1
python-decouple==3.8
python-jose==3.3.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
